Engine: Move SearchParameter from Core to Model namespace
Moves SearchParameter from Spark.Engine.Core to Spark.Engine.Model.
